// src/app/core/models/masters/audit-management/audit-statuses.ts
export interface AuditStatuses {
    id: number;
    title: string;
    type: string;
    status: string;
    status_id: number;
    status_label: string;
}

export interface AuditStatusesPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: AuditStatuses[];
}